requirements.txt
# Core Framework
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7

# Database
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
alembic==1.13.0

# Authentication & Security
Flask-Login==0.6.3
Flask-JWT-Extended==4.5.3
passlib==1.7.4
PyJWT==2.8.0
email-validator==2.1.0.post1

# Forms & Validation
WTForms==3.1.1
Flask-WTF==1.2.1

# Image Processing
Pillow==10.1.0

# Markdown Support
Markdown==3.5.1
bleach==6.1.0

# Email
Flask-Mail==0.9.1

# Development & Testing
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
flake8==6.1.0
black==23.11.0

# Production
gunicorn==21.2.0

# API & Cross-Origin
Flask-CORS==4.0.0
requests==2.31.0

# Utilities
python-slugify==8.0.1
PyYAML==6.0.1
humanize==4.9.0
pytz==2023.3.post1

# Rate Limiting
Flask-Limiter==3.5.0